adaptive_glass/requirements.txt
PyQt6
Pillow
piexif
